requirements.txt
yfinance>=0.2.36,<1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchainhub>=0.1.14
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
jinja2>=3.0.0
